double and int return double

+++ src/com/company/Main.java
@@ -3,6 +3,6 @@ package com.company;
 public class Main {
     public double sqr(double arg) { // надо double
         int k = 1;                  // есть int
-        return k;                   // можно неявно преобразовать int в double
+        return (double) k;          // явное преобразование int в double
     }
 }
